requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
yara-python>=4.5.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
prometheus-client>=0.20.0
jinja2>=3.1.0
pytest>=8.0.0
pytest-anyio>=0.0.0
anyio>=4.0.0
pytest-cov>=5.0.0
